requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
django-filter==23.5
drf-yasg==1.21.7
dj-database-url==2.1.0
gunicorn==21.2.0 